third_party/blink/web_tests/platform/generic/accessibility/aria-orientation-expected.txt

This tests that aria-orientation is reported correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS axObject.orientation is element.getAttribute('data-expected-orientation')
PASS successfullyParsed is true

TEST COMPLETE

50
50
50
50
50
50
50
50
Radio button
Radio button
Radio button
Button
Button
